handle delete image

+++ controllers/transactionController.js
@@ -98,9 +98,13 @@ class TransactionController {
             .then(data => {
                 if (data) {
                     receipt = data;
-                    let file = data.image_url.split('/');
-                    let fileName = file[file.length - 1];
-                    deleteFile(fileName);
+                    
+                    if(data.image_url) {
+                        let file = data.image_url.split('/');
+                        let fileName = file[file.length - 1];
+                        deleteFile(fileName);
+                    }
+                    
                     return User.findById(req.decode._id)
                 } else {
                     res.status(404).json({ message: `cant find bill with id : ${req.params.id}` });
